requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
async-timeout==4.0.2
    # via redis
brotli==1.0.9
    # via flask-compress
click==8.1.3
    # via flask
commonmark==0.9.1
    # via rich
cssmin==0.2.0
    # via -r requirements.in
deepmerge==1.1.0
    # via govuk-frontend-wtf
deprecated==1.2.13
    # via limits
dnspython==2.2.1
    # via email-validator
email-validator==1.3.0
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-assets
    #   flask-compress
    #   flask-limiter
    #   flask-wtf
    #   govuk-frontend-wtf
flask-assets==2.0
    # via -r requirements.in
flask-compress==1.13
    # via -r requirements.in
flask-limiter==3.1.0
    # via -r requirements.in
flask-talisman==1.0.0
    # via -r requirements.in
flask-wtf==1.0.1
    # via govuk-frontend-wtf
govuk-frontend-jinja==2.4.0
    # via
    #   -r requirements.in
    #   govuk-frontend-wtf
govuk-frontend-wtf==2.2.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via email-validator
importlib-metadata==6.0.0
    # via flask
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via
    #   flask
    #   govuk-frontend-jinja
    #   govuk-frontend-wtf
jsmin==3.0.1
    # via -r requirements.in
limits==3.1.5
    # via flask-limiter
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
ordered-set==4.1.0
    # via flask-limiter
packaging==23.0
    # via limits
pygments==2.14.0
    # via rich
python-dotenv==0.21.0
    # via -r requirements.in
redis==4.4.2
    # via -r requirements.in
rich==12.6.0
    # via flask-limiter
typing-extensions==4.4.0
    # via
    #   flask-limiter
    #   limits
webassets==2.0
    # via flask-assets
werkzeug==2.2.2
    # via flask
wrapt==1.14.1
    # via deprecated
wtforms==3.0.1
    # via
    #   flask-wtf
    #   govuk-frontend-wtf
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
async-timeout==4.0.2
attrs==22.2.0
bandit==1.7.4
black==22.12.0
bleach==6.0.0
bokeh==3.0.3
Brotli==1.0.9
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
coverage==7.0.5
cssmin==0.2.0
db-dtypes==1.0.5
deepmerge==1.1.0
Deprecated==1.2.13
dnspython==2.2.1
docutils==0.19
dparse==0.6.2
email-validator==1.3.0
exceptiongroup==1.1.0
flake8==4.0.1
flake8-bugbear==22.12.6
Flask==2.2.2
Flask-Assets==2.0
Flask-Compress==1.13
Flask-Limiter==3.1.0
flask-talisman==1.0.0
Flask-WTF==1.0.1
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
govuk-frontend-jinja==2.4.0
govuk-frontend-wtf==2.2.0
grpcio==1.51.1
grpcio-status==1.51.1
gspread==5.7.2
gunicorn==20.1.0
hacking==5.0.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.11.4
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsmin==3.0.1
keyring==23.13.1
limits==3.1.5
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==9.0.0
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
ordered-set==4.1.0
packaging==21.3
pandas==1.5.3
pathspec==0.10.3
pbr==5.11.1
pep8-naming==0.13.2
Pillow==9.4.0
pip-tools==6.12.1
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.21.12
pur==7.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.14.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
readme-renderer==37.3
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safety==2.3.5
six==1.16.0
smmap==5.0.0
stevedore==4.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
twine==4.0.2
typing_extensions==4.4.0
urllib3==1.26.14
webassets==2.0
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xyzservices==2022.9.0
zipp==3.11.0
